cv_process/requirements.txt
absl-py==1.4.0
anyio==4.3.0
apache-beam==2.46.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array-record==0.4.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
avro-python3==1.10.2
Babel==2.15.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
comm==0.2.2
contextlib2==21.6.0
contourpy==1.1.1
crcmod==1.7
cycler==0.12.1
Cython==3.0.10
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.1.1
dm-tree==0.1.8
docopt==0.6.2
easyocr==1.7.1
editdistance==0.8.1
efficientnet==1.0.0
essential-generators==1.0
etils==1.3.0
exceptiongroup==1.2.1
executing==2.0.1
fastavro==1.9.4
fasteners==0.19
fastjsonschema==2.19.1
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.52.1
fqdn==1.5.1
fsspec==2024.5.0
gast==0.4.0
gin-config==0.5.0
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.63.0
graphviz==0.20.3
grpcio==1.64.0
h11==0.14.0
h5py==3.11.0
hdfs==2.7.3
httpcore==1.0.5
httplib2==0.21.0
httpx==0.27.0
idna==3.7
imageio==2.34.1
imgaug==0.4.0
immutabledict==4.2.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
inline==0.0.1
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.12.3
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.10
kaggle==1.6.14
keras==2.10.0
Keras-Applications==1.0.8
keras-ocr==0.9.3
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
lvis==0.5.3
lxml==5.2.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mistune==3.0.2
mkl==2021.4.0
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.1
notebook==7.2.0
notebook_shim==0.2.4
numpy==1.24.4
oauth2client==4.1.3
oauthlib==3.2.2
objsize==0.6.1
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
orjson==3.10.3
overrides==7.7.0
packaging==24.0
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.3.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
portalocker==2.8.2
prometheus_client==0.20.0
promise==2.3
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==3.19.6
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pycocotools==2.0.7
pycparser==2.22
pydot==1.4.2
Pygments==2.18.0
pymongo==3.13.0
pyparsing==2.4.7
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-slugify==8.0.4
pytz==2024.1
PyWavelets==1.4.1
pywin32==306
pywinpty==2.0.13
PyYAML==5.4.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.2
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.1
rsa==4.9
sacrebleu==2.2.0
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
Send2Trash==1.8.3
sentencepiece==0.2.0
seqeval==1.2.2
shapely==2.0.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tbb==2021.12.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-addons==0.21.0
tensorflow-datasets==4.9.0
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.0
tensorflow-hub==0.16.1
tensorflow-io==0.31.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.13.0
tensorflow-model-optimization==0.8.0
tensorflow-text==2.10.0
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
tf-keras==2.15.0
tf-models-official==2.10.1
tf-slim==1.1.0
threadpoolctl==3.5.0
tifffile==2023.7.10
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typeguard==2.13.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.0
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
validators==0.28.3
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wget==3.2
widgetsnbextension==4.0.10
wrapt==1.16.0
zipp==3.18.2
zstandard==0.22.0